% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{data}
\name{amrn}
\alias{amrn}
\title{Arabidopsis morphogenesis regulatory network.}
\format{A data frame with 22 rows and 3 variables:
\describe{
  \item{\code{Source}}{the identifier of the source node}
  \item{\code{Interaction}}{interaction type of the edge}
  \item{\code{Target}}{the identifier of the target node}
  ...
}}
\source{
\url{http://www.ncbi.nlm.nih.gov/pubmed/10487867}
}
\usage{
data(amrn)
}
\description{
The Arabidopsis morphogenesis regulatory network (AMRN) with 10 nodes and 22 links.
This regulatory network is known to robustly control the process of flower development.
}
\keyword{datasets}

