% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{calCentrality}
\alias{calCentrality}
\title{Calculate node-/edge-based centralities of a network or a set of networks.}
\usage{
calCentrality(networks)
}
\arguments{
\item{networks}{A network or a set of networks used for the calculation}
}
\value{
The updated network objects including values of centrality measures for each node/edge.
}
\description{
Calculate centrality measures for all nodes/edges in a network or in a set of networks.
}
\details{
This function calculates node-/edge-based centralities of a specific network or a set of networks.
For each network, the returned results are stored in a data frame of the network object.
The data frame has one column for nodes/edges identifiers, and nine columns
contain corresponding values of some node-based centrality measures such as
Degree, In-/Out-Degree, Closeness, Betweenness, Stress and Eigenvector, and some edge-based measures like
Edge Degree and Edge Betweenness.
}
\examples{
data(amrn)
amrn <- calCentrality(amrn)
print(amrn$nodes)
print(amrn$edges)
}
\seealso{
\code{\link{findFBLs}}, \code{\link{findFFLs}}, \code{\link{calSensitivity}}
}

