% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{data}
\name{cchs}
\alias{cchs}
\title{Cell cycle pathway of the species Homo sapiens.}
\format{A data frame with 223 rows and 3 variables:
\describe{
  \item{\code{Source}}{the identifier of the source node}
  \item{\code{Interaction}}{interaction type of the edge}
  \item{\code{Target}}{the identifier of the target node}
  ...
}}
\source{
\url{https://www.ncbi.nlm.nih.gov/pubmed/9733515}
}
\usage{
data(cchs)
}
\description{
The cell cycle pathway of the species Homo sapiens (CCHS) with 161 nodes and 223 links.
The cell cycle is the series of events that takes place in a cell
leading to its division and duplication (replication).
}
\keyword{datasets}
