% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{data}
\name{ccsn}
\alias{ccsn}
\title{Canonical cell signaling network.}
\format{A data frame with 1633 rows and 3 variables:
\describe{
  \item{\code{Source}}{the identifier of the source node}
  \item{\code{Interaction}}{interaction type of the edge}
  \item{\code{Target}}{the identifier of the target node}
  ...
}}
\source{
\url{http://bioinformatics.oxfordjournals.org/content/24/17/1926.long}
}
\usage{
data(ccsn)
}
\description{
The canonical cell signaling network (CCSN) with 771 nodes and 1633 links.
The network was obtained from \url{http://stke.sciencemag.org/}, and all the neutral interactions were excluded.
}
\keyword{datasets}
