% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{data}
\name{cdrn}
\alias{cdrn}
\title{Cell differentiation regulatory network.}
\format{A data frame with 15 rows and 3 variables:
\describe{
  \item{\code{Source}}{the identifier of the source node}
  \item{\code{Interaction}}{interaction type of the edge}
  \item{\code{Target}}{the identifier of the target node}
  ...
}}
\source{
\url{http://www.ncbi.nlm.nih.gov/pubmed/11082279}
}
\usage{
data(cdrn)
}
\description{
The cell differentiation regulatory network (CDRN) with 9 nodes and 15 links.
CDRN has seven positive and two negative FBLs is found to robustly
induce quiescence, terminal differentiation, and apoptosis.
}
\keyword{datasets}

