% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{findAttractors}
\alias{findAttractors}
\title{Identifies attractors of a network.}
\usage{
findAttractors(network, stateSet)
}
\arguments{
\item{network}{A network used for the attractors search}

\item{stateSet}{The identifier for accessing a set of initial-states}
}
\value{
The resulted transition network.
}
\description{
Identifies attractors of a network and returns the resulted transition network.
}
\details{
This function searches attractors of a specific network, and
the returned results are stored in a Transition network object.
The calculation is based on a set of initial-states specified by an identifier \strong{\code{stateSet}}.
The current set of update-rules of the network is used with a synchronous updating scheme.
}
\examples{
data(amrn)
# Generate a set of random Nested Canalyzing rules
generateRule(amrn, 2)

# Generate a specific initial-state for the AMRN network
state1 <- generateState(amrn, "1110011011")

transNet <- findAttractors(amrn, state1)
print(transNet)
output(transNet)
}
\seealso{
\code{\link{generateStates}}, \code{\link{generateState}}, \code{\link{generateRule}}, \code{\link{calSensitivity}}
}
