% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{findFBLs}
\alias{findFBLs}
\title{Identifies feedback loops in a network or in a set of networks.}
\usage{
findFBLs(networks, maxLength = 2L)
}
\arguments{
\item{networks}{A network or a set of networks used for FBLs search}

\item{maxLength}{The maximal length of FBLs, default is 2}
}
\value{
The updated network objects including number of FBLs for each node/edge.
}
\description{
Searches and counts feedback loops for all nodes/edges in a network or in a set of networks.
}
\details{
This function searches feedback loops (FBLs) in a specific network or in a set of networks.
For each network, the returned results are stored in two corresponding data frames of node/edge attributes.
Each data frame has one column for node/edge identifiers, and three columns
contain corresponding number of involved FBLs and number of involved positive/negative FBLs.
Another data frame of the network object, "network" data frame,
contains total number of FBLs and total number of positive/negative FBLs in the network.
}
\examples{
data(amrn)
amrn <- findFBLs(amrn, maxLength = 10)
print(amrn$nodes)
print(amrn$edges)
print(amrn$network)
}
\seealso{
\code{\link{findFFLs}}, \code{\link{calCentrality}}, \code{\link{calSensitivity}}
}
