% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{findFFLs}
\alias{findFFLs}
\title{Identifies feedforward loops in a network or in a set of networks.}
\usage{
findFFLs(networks)
}
\arguments{
\item{networks}{A network or a set of networks used for FFLs search}
}
\value{
The updated network objects including number of FFL motifs for each node/edge.
}
\description{
Searches and counts feedforward loops for all nodes/edges in a network or in a set of networks.
}
\details{
This function searches feedforward loops (FFLs) in a specific network or in a set of networks.
For each network, the returned results are stored in two corresponding data frames of node/edge attributes.
Each data frame has one column for node/edge identifiers, and four columns
contain corresponding number of all FFL motifs and number of FFL motifs
with three different roles A, B and C.
Another data frame of the network object, "network" data frame, contains total numbers of FFLs and coherent/incoherent FFLs in the network.
}
\examples{
data(amrn)
amrn <- findFFLs(amrn)
print(amrn$nodes)
print(amrn$edges)
print(amrn$network)
}
\seealso{
\code{\link{findFBLs}}, \code{\link{calCentrality}}, \code{\link{calSensitivity}}
}

