% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{generateGroup}
\alias{generateGroup}
\title{Generate a specific group of nodes/edges.}
\usage{
generateGroup(network, nodes, edges = "")
}
\arguments{
\item{network}{The network contains the generated group}

\item{nodes}{A list of nodes in the generated group}

\item{edges}{A list of edges in the generated group}
}
\value{
The updated network object including the generated group
}
\description{
Generate a specific group of nodes/edges in a network.
}
\details{
This function generates a specific group of elements in a network.
The group would be used to analyze the dynamics of the examined network,
for ex., calculating sensitivity,  perturbing the network, or restoring the network to the origin.
The element group contains only nodes, only edges, or a combination of nodes/edges.
}
\examples{
data(amrn)
# Generate a group of two nodes and two edges in the AMRN network
amrn <- generateGroup(amrn, nodes = "AG, SUP", edges = "UFO (1) PI, LUG (-1) PI")
print(amrn$Group_1)
}
\seealso{
\code{\link{calSensitivity}}, \code{\link{perturb}}, \code{\link{restore}}
}

