% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{generateRule}
\alias{generateRule}
\title{Generate a default set of update-rules.}
\usage{
generateRule(network, ruleType = 0L)
}
\arguments{
\item{network}{The network used for the generation}

\item{ruleType}{Type of random update-rules, default is 0}
}
\value{
The string "ok" if success, otherwise NULL object
}
\description{
Generate a default set of update-rules for a network.
}
\details{
This function generates a default set of update-rules for a network.
The rules would be used to analyze the dynamics of the examined network, for ex., calculating sensitivity, searching attractors.
The type of random update-rules can be specified by the parameter \code{ruleType}:
0 means only Conjunction rules, 1 means only Disjunction rules and 2 means random Nested Canalyzing rules.
}
\examples{
data(amrn)
# Generate a set of random Nested Canalyzing rules
generateRule(amrn, 2)

# Generate a specific initial-state for the AMRN network
state1 <- generateState(amrn, "1110011011")

att <- findAttractors(amrn, state1)
print(att)
}
\seealso{
\code{\link{findAttractors}}, \code{\link{perturb}}, \code{\link{restore}}
}
