% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{generateState}
\alias{generateState}
\title{Generate a specific initial-state.}
\usage{
generateState(network, state)
}
\arguments{
\item{network}{The network used for the generation}

\item{state}{A binary string with one entry for each node of the network in alphabetical order}
}
\value{
An identifier for accessing the generated initial-state. The identifier would be used as a parameter of the functions of calculating sensitivity and finding attractors.
}
\description{
Generate a specific initial-state for a network.
}
\details{
This function generates a specific initial-state for a network.
The initial-state would be used to analyze the dynamics of the examined network,
for ex., calculating sensitivity or searching attractors.
}
\examples{
data(amrn)
state1 <- generateState(amrn, "1010101111")
print(state1)
}
\seealso{
\code{\link{calSensitivity}}, \code{\link{findAttractors}}
}
