% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{generateStates}
\alias{generateStates}
\title{Generate random initial-states.}
\usage{
generateStates(numNodes, numStates)
}
\arguments{
\item{numNodes}{Number of nodes in each initial-state}

\item{numStates}{Number of random initial-states to be generated. If set to "all", all possible initial-states would be generated. For the large networks, we should use a specific value becaused of memory limitation.}
}
\value{
An identifier for accessing the generated initial-states. The identifier would be used as a parameter of the functions of calculating sensitivity and finding attractors.
}
\description{
Generate random initial-states for a network or a set of networks.
}
\details{
This function generates random initial-states for a network or a set of networks.
The initial-states would be used to analyze the dynamics of the examined networks,
for ex., calculating sensitivity or searching attractors.
}
\examples{
# Generate a set of 200 initial-states each containing 10 Boolean nodes
set1 <- generateStates(10, 200)
print(set1)

# Generate all possible initial-states each containing 10 Boolean nodes
set2 <- generateStates(10, "all")
print(set2)
}
\seealso{
\code{\link{calSensitivity}}, \code{\link{findAttractors}}
}

