% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{data}
\name{hsn}
\alias{hsn}
\title{Human signaling network.}
\format{A data frame with 3102 rows and 3 variables:
\describe{
  \item{\code{Source}}{the identifier of the source node}
  \item{\code{Interaction}}{interaction type of the edge}
  \item{\code{Target}}{the identifier of the target node}
  ...
}}
\source{
\url{http://www.ncbi.nlm.nih.gov/pubmed/19226461}
}
\usage{
data(hsn)
}
\description{
The large-scale human signaling network (HSN) with 1192 nodes and 3102 links.
Based on the network, some general principles were provided for understanding protein evolution in the context of signaling networks.
}
\keyword{datasets}
