% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{loadNetwork}
\alias{loadNetwork}
\title{Loads a network from a file.}
\usage{
loadNetwork(pathToFile)
}
\arguments{
\item{pathToFile}{The path points to a file}
}
\value{
The network object
}
\description{
\code{loadNetwork} loads a network from a file and returns the network object.
}
\details{
This function loads a network from a Tab-separated values text file and returns the network object.
The file format contains three columns:  source, interaction type, and target.
"Source" and "target" are gene/protein identifiers that are used to define nodes,
while "interaction type" labels the edges connecting each pair of nodes.
The returned network object contains the network name, three data frames used for storing
the nodes/edges and network attributes, respectively.
}
\examples{
amrn <- loadNetwork("D:\\\\AMRN.sif")
print(amrn)
}
\seealso{
\code{\link{output}}, \code{\link{createRBNs}}, \code{\link{calSensitivity}}
}
