% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{output}
\alias{output}
\title{Exports all node/edge/network attributes of a network or a set of networks.}
\usage{
output(networks)
}
\arguments{
\item{networks}{The network or the set of networks}
}
\description{
\code{output} writes all node/edge/network attributes of a network or a set of networks into CSV files.
}
\details{
This function writes all node/edge/network attributes of a network or a set of networks into CSV files.
For each network, the function exports all data frames of the network object containing
structural attributes of the nodes/edges/network and sensitivity values of mutated groups.

The CSV files were outputed with names as follows: \strong{\code{[network name]}_out_\code{[data-frame name]}.csv}.
The structure of these networks are also exported as Tab-separated values text files (.SIF extension).
}
\examples{
data(amrn)
# Generate all possible initial-states each containing 10 Boolean nodes
set1 <- generateStates(10, "all")

# Generate all possible groups each containing a single node in the AMRN network
amrn <- generateGroups(amrn, "all", 1, 0)
amrn <- calSensitivity(amrn, set1, 1, "knockout")

output(amrn)
}
\seealso{
\code{\link{loadNetwork}}, \code{\link{createRBNs}}
}

