% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{printSensitivity}
\alias{printSensitivity}
\title{Print out the sensitivity values of node/edge groups}
\usage{
printSensitivity(network, groupSet = 0)
}
\arguments{
\item{network}{A network used for the outputting}

\item{groupSet}{The indexing number of node/edge groups for whose sensitivity values are calculated. Default is 0 which specify the latest generated groups.}
}
\value{
None
}
\description{
Print out the sensitivity values of node/edge groups in a network
}
\details{
This function prints out the sensitivity values of node/edge groups in a specific network.
And the parameter \code{groupSet}
has same meaning as in the \code{\link{calSensitivity}} function.
}
\examples{
data(amrn)

# generate 1000 random initial-states
states <- generateStates(amrn, 1000)
print(states)

# generate all possible groups each containing a single node in the HSN network
amrn <- generateGroups(amrn, "all", 1, 0)

# calculate sensitivity values of all nodes against the knockout mutation
amrn <- calSensitivity(amrn, states, "knockout")

# view the calculated sensitivity values and export all results to files
printSensitivity(amrn)
}
\seealso{
\code{\link{calSensitivity}}, \code{\link{generateStates}}, \code{\link{generateGroups}}, \code{\link{generateGroup}}, \code{\link{findFBLs}}, \code{\link{findFFLs}}, \code{\link{calCentrality}}, \code{\link{findAttractors}}
}
