% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{setOpencl}
\alias{setOpencl}
\title{Enables or disables OpenCL computation.}
\usage{
setOpencl(deviceType)
}
\arguments{
\item{deviceType}{The type of OpenCL device, including three options: \'none\', \'cpu\' and \'gpu\'}
}
\value{
Information of the successfully selected device
}
\description{
\code{setOpencl} enables or disables OpenCL computation.
}
\details{
This function enables OpenCL computation by selecting a CPU/GPU device and utilizing all of its cores for further computation.
Thus, all tasks will be executed in parallel.
About the parameter \code{deviceType}, there exists three options: \'none\' means disable OpenCL,
\'cpu\' means selecting a CPU device and \'gpu\' means using a GPU device.
}
\examples{
showOpencl()
setOpencl("gpu")
setOpencl("cpu")
setOpencl("none")

}
\seealso{
\code{\link{showOpencl}}
}
