% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{showOpencl}
\alias{showOpencl}
\title{Shows OpenCL information.}
\usage{
showOpencl()
}
\value{
A string of OpenCL information
}
\description{
\code{showOpencl} gets OpenCL information and prints them to the console screen.
}
\details{
This function gets OpenCL information and prints them to the console screen.
For ex., installed OpenCL platforms and its corresponding CPU/GPU devices.
}
\examples{
showOpencl()

}
\seealso{
\code{\link{setOpencl}}
}

